+++ apps/demo/emails/yelp-recent-login.tsx
@@ -105,9 +105,7 @@ export const YelpRecentLoginEmail = ({
             </Row>
             <Row style={{ ...boxInfos, paddingTop: '0' }}>
               <Column style={containerButton} colSpan={2}>
-                <Button pY={12} pX={30} style={button}>
-                  Learn More
-                </Button>
+                <Button style={button}>Learn More</Button>
               </Column>
             </Row>
           </Section>
@@ -167,7 +165,8 @@ const button = {
   color: '#FFF',
   fontWeight: 'bold',
   border: '1px solid rgb(0,0,0, 0.1)',
-  cursor: 'pointer'
+  cursor: 'pointer',
+  padding: '12px 30px'
 };
 
 const content = {