+++ apps/demo/staged/netlify-welcome.tsx
@@ -119,7 +119,7 @@ export const NetlifyWelcomeEmail = ({
             <ul>{steps?.map(({ Description }) => Description)}</ul>
 
             <Section className="text-center">
-              <Button className="bg-brand text-white rounded-lg" pY={12} pX={18}>
+              <Button className="bg-brand text-white rounded-lg py-[18px] px-3">
                 Go to your dashboard
               </Button>
             </Section>