+++ packages/cli/app/vite.config.ts
@@ -1,16 +1,22 @@
 import react from '@vitejs/plugin-react';
 import { defineConfig } from 'vite';
-import dynamicImport from 'vite-plugin-dynamic-import';
 
 process.chdir(__dirname);
 
 export default defineConfig({
-  build: {
-    target: 'esnext'
-  },
   define: {
     'process.env': 'import.meta.env'
   },
-  plugins: [dynamicImport(), react()],
+  optimizeDeps: {
+    include: [
+      'classnames',
+      'deepmerge',
+      'pretty',
+      'react-dom',
+      'react-dom/client',
+      'react-dom/server'
+    ]
+  },
+  plugins: [react()],
   root: __dirname
 });