+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,92 @@
 # @jsx-email/cli ChangeLog
 
+## v1.0.9
+
+_2023-10-04_
+
+### Bugfixes
+
+- fix: use vite builtins for perf improvement in preview (#28)
+- fix: add shebang to cli script (#30)
+- fix: typos on cli and documentation (#29)
+- fix: rework the dynamic imports for the preview tool (#27)
+- fix: use @jsx-email/all in template (f47a5d5)
+
+## v1.0.8
+
+_2023-10-04_
+
+### Bugfixes
+
+- fix: add shebang to cli script (#30)
+- fix: typos on cli and documentation (#29)
+- fix: rework the dynamic imports for the preview tool (#27)
+- fix: use @jsx-email/all in template (f47a5d5)
+- fix: mustache template path, character escaping (#25)
+
+## v1.0.7
+
+_2023-10-04_
+
+### Bugfixes
+
+- fix: typos on cli and documentation (#29)
+- fix: rework the dynamic imports for the preview tool (#27)
+- fix: use @jsx-email/all in template (f47a5d5)
+- fix: mustache template path, character escaping (#25)
+
+## v1.0.6
+
+_2023-10-04_
+
+### Bugfixes
+
+- fix: rework the dynamic imports for the preview tool (#27)
+- fix: use @jsx-email/all in template (f47a5d5)
+- fix: mustache template path, character escaping (#25)
+
+## v1.0.5
+
+_2023-10-03_
+
+### Bugfixes
+
+- fix: use @jsx-email/all in template (f47a5d5)
+- fix: mustache template path, character escaping (#25)
+
+## v1.0.4
+
+_2023-10-03_
+
+### Bugfixes
+
+- fix: mustache template path, character escaping (#25)
+
+## v1.0.3
+
+_2023-10-03_
+
+### Bugfixes
+
+- fix: add create command to main help message (b82a85b)
+
+## v1.0.2
+
+_2023-10-02_
+
+### Bugfixes
+
+- fix: add create command to main help message (b82a85b)
+- fix: pick up secondary changes (faaade5)
+
+## v1.0.1
+
+_2023-10-02_
+
+### Bugfixes
+
+- fix: pick up secondary changes (faaade5)
+
 ## v1.0.0
 
 _2023-09-27_