+++ packages/font/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jsx-email/font ChangeLog
 
+## v1.0.1
+
+_2023-10-02_
+
+### Bugfixes
+
+- fix: add more robust fallback stack. credit @boredland (d48a4c4)
+
 ## v1.0.0
 
 _2023-09-27_